requirements.txt
groq
google-genai
requests
python-dotenv
cohere
huggingface_hub
ollama